requirements.txt
absl-py==2.1.0
asgiref==3.8.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
djangorestframework==3.15.0
filelock==3.13.2
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pydicom==2.4.4
Pygments==2.17.2
pyjpegls==1.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.23
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.34
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
